<feat!> finished admin part; finished delete meals feature; added several related methods in `Controller`;
<fix> revised several Model methods; revised several javadoc; revised the `addBanquet()` logic where `Available` should always be `N`; revised database test data

+++ src/service/managers/MealsManager.java
@@ -262,7 +262,7 @@ public final class MealsManager {
      * @param ID the uniqure identifier of the meal to delete
      * @throws ModelException if any errors encountered.
      */
-    public void deleteMeal(int BIN, int ID) throws ModelException {
+    public void deleteMeal(long BIN, long ID) throws ModelException {
         String stmt = "DELETE FROM MEALS WHERE BIN = ? AND ID = ?";
 
         try (PreparedStatement pstmt = con.getConnection().prepareStatement(stmt)) {
